integration: Move storage helper functions to `testing`
